Moved all teleOp Configs into package
