+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/configs/teleOpInterface.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.configs;
 
 import java.util.HashMap;
